python-backend/requirements.txt
blinker==1.9.0
click==8.2.1
dnspython==2.7.0
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mongoengine==0.29.1
pymongo==4.13.1
python-dotenv==1.1.0
redis==6.2.0
Werkzeug==3.1.3
